Move http modules to src/http and update import paths
