+++ portal/common/lib/tests/resource.test.ts
@@ -5,7 +5,7 @@
 import { beforeEach, describe, expect, test, vi } from "vitest";
 import { ResourceFetcher } from "../src/resource";
 import { RPCSelector } from "../src/rpc_selector";
-import { HttpStatusCodes } from "../http/http_status_codes";
+import { HttpStatusCodes } from "../src/http/http_status_codes";
 import { checkRedirect } from "../src/redirects";
 import { fromBase64 } from "@mysten/bcs";
 import { SuiObjectResponse } from "@mysten/sui/client";
